requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-groq
langchain-community
pillow
numpy
pandas
scikit-learn
tensorflow